reduce code mess by shifting relavant sections to appropriate helpers
